Add files via upload
To convert the information inside the TXT file to the Excel file, the format of the TXT file must be determined first. We assume that the TXT file contains data with a certain format, for example, the data is separated by commas (,) or hyphens (-).

Next, a Python source code example for reading data from a TXT file and writing them to an Excel file is given. This code uses pandas and openpyxl libraries.

By running this code, the data in the data.txt file will be converted to the data.xlsx file and you will have an Excel file with the same data. Make sure the data.txt file is located in the same path as your Python code is running, or specify the exact file path in the file_path variable.